--- a/mujoco_warp/test_data/actuation/actuators.xml
+++ b/mujoco_warp/test_data/actuation/actuators.xml
@@ -16,10 +16,13 @@
     <!-- dyntype -->
     <general joint="hinge" dyntype="none"/>
     <general joint="hinge" dyntype="integrator"/>
+    <general joint="hinge" dyntype="integrator" actearly="true"/>
     <general joint="hinge" dyntype="filter"/>
+    <general joint="hinge" dyntype="filter" actearly="true"/>
     <general joint="hinge" dyntype="filterexact"/>
+    <general joint="hinge" dyntype="filterexact" actearly="true"/>
   </actuator>
   <keyframe>
-    <key ctrl="1 1 1 1 1 1 1 1" act=".1 .2 .3" qpos=".1234" qvel=".2345"/>
+    <key ctrl="1 1 1 1 1 1 1 1 1 1 1" act=".1 .1 .2 .2 .3 .3" qpos=".1234" qvel=".2345"/>
   </keyframe>
 </mujoco>