https://mujoco.readthedocs.io/en/latest/XMLreference.html#actuator-general-actearly