@thowell 2025-07-03
to test an environment with activation states the mujoco menagerie aloha environment (https://github.com/google-deepmind/mujoco_menagerie/tree/main/aloha) is modified to have filtered cartesian actuators (https://github.com/google-deepmind/mujoco_menagerie/blob/main/aloha/filtered_cartesian_actuators.xml)

**aloha + cartesian filtered actuators**
```
mjwarp-testspeed --function=step --mjcf=mujoco_menagerie/aloha/scene.xml --batch_size=8192 --nconmax=2560000 --njmax=2560000
```

this pr:
```
Summary for 8192 parallel rollouts

 Total JIT time: 1.05 s
 Total simulation time: 302.93 s
 Total steps per second: 27,043
 Total realtime factor: 54.09 x
 Total time per step: 36978.46 ns
```

main:
```
Summary for 8192 parallel rollouts

 Total JIT time: 1.06 s
 Total simulation time: 302.21 s
 Total steps per second: 27,107
 Total realtime factor: 54.21 x
 Total time per step: 36890.85 ns
```

**humanoid**
```
mjwarp-testspeed --function=step --mjcf=mujoco_warp/mujoco_warp/test_data/humanoid/humanoid.xml --batch_size=8192 --nconmax=512000 --njmax=512000
```

this pr:
```
Summary for 8192 parallel rollouts

 Total JIT time: 0.28 s
 Total simulation time: 30.14 s
 Total steps per second: 271,811
 Total realtime factor: 1,359.05 x
 Total time per step: 3679.03 ns
```

main:
```
Summary for 8192 parallel rollouts

 Total JIT time: 0.29 s
 Total simulation time: 30.22 s
 Total steps per second: 271,120
 Total realtime factor: 1,355.60 x
 Total time per step: 3688.40 ns
```

**summary: does not look like `actearly` has significant impact on overall performance**